+++ src/components/fields/learning-goal-field.tsx
@@ -1,8 +1,8 @@
+import type { ControlledFieldProps } from './types'
+import { type FieldValues, type Path, useController } from 'react-hook-form'
 import { RadioGroup, RadioGroupItem } from '@/components/ui/radio-group'
 import { Label } from '@/components/ui/label'
 import { Users, GraduationCap, Briefcase } from 'lucide-react'
-import { ControlledFieldProps } from '.'
-import { Path, useController, FieldValues } from 'react-hook-form'
 import { ShowError } from '@/components/errors'
 import { cn } from '@/lib/utils'
