+++ src/routes/_user/learn.$lang.add-phrase.tsx
@@ -1,3 +1,4 @@
+import { useCallback } from 'react'
 import { createFileRoute, Link } from '@tanstack/react-router'
 import { useMutation } from '@tanstack/react-query'
 import { Controller, SubmitHandler, useForm } from 'react-hook-form'
@@ -19,12 +20,9 @@ import { NotebookPen, Search } from 'lucide-react'
 import languages from '@/lib/languages'
 import { Loader } from '@/components/ui/loader'
 import supabase from '@/lib/supabase-client'
-import {
-	TranslationLanguageField,
-	TranslationTextField,
-} from '@/components/fields'
-import { useCallback } from 'react'
 import { PhraseCardInsert } from '@/types/main'
+import TranslationTextField from '@/components/fields/translation-text-field'
+import TranslationLanguageField from '@/components/fields/translation-language-field'
 
 interface SearchParams {
 	text?: string
