+++ src/routes/_user/learn.add-deck.tsx
@@ -2,19 +2,19 @@ import { createFileRoute, Link } from '@tanstack/react-router'
 import { SubmitHandler, useController, useForm } from 'react-hook-form'
 import { z } from 'zod'
 import { zodResolver } from '@hookform/resolvers/zod'
-import { ShowAndLogError } from '@/components/errors'
-import { SelectOneLanguage } from '@/components/select-one-language'
+import { RouteIcon } from 'lucide-react'
+import type { TitleBar } from '@/types/main'
+import { Badge } from '@/components/ui/badge'
 import { Button } from '@/components/ui/button'
 import { buttonVariants } from '@/components/ui/button-variants'
+import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
+import Callout from '@/components/ui/callout'
+import { ShowAndLogError } from '@/components/errors'
+import { SelectOneLanguage } from '@/components/select-one-language'
 import { useNewDeckMutation } from '@/lib/mutate-deck'
 import { useProfile } from '@/lib/use-profile'
-import { TitleBar } from '@/types/main'
-import { Badge } from '@/components/ui/badge'
 import languages from '@/lib/languages'
-import Callout from '@/components/ui/callout'
-import { ErrorLabel } from '@/components/fields'
-import { RouteIcon } from 'lucide-react'
-import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
+import ErrorLabel from '@/components/fields/error-label'
 
 const SearchSchema = z.object({
 	lang: z.string().optional(),
